requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-WTF==1.1.1
Flask-Talisman==1.0.0
Werkzeug==2.3.7
SQLAlchemy==2.0.21
WTForms==3.0.1
python-dotenv==1.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.7
redis==5.0.1
celery==5.3.4
prometheus-client==0.17.1
psutil==5.9.6
sentry-sdk[flask]==1.32.0
certbot==2.6.0
python-dateutil==2.8.2
numpy==1.24.3
scikit-learn==1.3.0
ruptures==1.1.8
hmmlearn==0.3.0
pandas==2.0.3
scipy==1.11.32
joblib==1.3.2
matplotlib==3.8.1
lightgbm==4.4.0
python-dotenv==1.0.0